chore: update chokidar to v5
Actual change is just the update, but because chokidar is ESM-only, jest immideately died.
So I had to move out `isIgnoredByWatcher` function which is used in `watcher-ignored-test.ts`.
Now, `config.ts` is not opened by jest and we are happy.

Refs: #14991
