+++ packages/react-router-dev/config/config.ts
@@ -22,6 +22,7 @@ import {
   configRoutesToRouteManifest,
 } from "./routes";
 import { detectPackageManager } from "../cli/detectPackageManager";
+import { isIgnoredByWatcher } from "./is-ignored-by-watcher";
 
 const excludedConfigPresetKeys = ["presets"] as const satisfies ReadonlyArray<
   keyof ReactRouterConfig
@@ -1159,38 +1160,6 @@ function isEntryFileDependency(
     ) {
       return true;
     }
-  }
-
-  return false;
-}
-
-export function isIgnoredByWatcher(
-  path: string,
-  { root, appDirectory }: { root: string; appDirectory: string },
-): boolean {
-  let dirname = Path.dirname(path);
-
-  let ignoredByPath =
-    !dirname.startsWith(appDirectory) &&
-    // Ensure we're only watching files outside of the app directory
-    // that are at the root level, not nested in subdirectories
-    path !== root && // Watch the root directory itself
-    dirname !== root; // Watch files at the root level
-
-  if (ignoredByPath) {
-    return true;
-  }
-
-  // Filter out non-regular files (sockets, pipes, etc.) that
-  // crash `fs.watch()` on macOS with errno -102
-  // https://github.com/paulmillr/chokidar/issues/1391
-  try {
-    let stat = fs.statSync(path, { throwIfNoEntry: false });
-    if (stat && !stat.isFile() && !stat.isDirectory()) {
-      return true;
-    }
-  } catch {
-    return true;
   }
 
   return false;
